Module bittongue::grapheme

This module exports utilites related to grapheme clusters/unicode segmentation.

Some “visual” characteres are actually multiple unicode characters, such as letter “a” with three diacritics “ā̤́”, those are grapheme clusters.

Structs

GraphemeCluster

A single grapheme cluster. Unsized type.

IndexedIter

An indexed iterator over grapheme clusters of a string, yielding grapheme indices in terms of bytes together with the graphemes.

Iter

An iterator over grapheme clusters of a string.

Functions

indexed_iter

Creates an indexed iterator over grapheme clusters of a string, yielding grapheme indices in terms of bytes together with the graphemes.

iter

Creates an iterator over grapheme clusters of a string.