Struct bittongue::grapheme::GraphemeCluster

#[repr(transparent)]
pub struct GraphemeCluster { /* fields omitted */ }

A single grapheme cluster. Unsized type.

Implementations

impl GraphemeCluster

pub fn new(content: &str) -> Option<&Self>

Creates a new grapheme cluster reference from a string borrowed reference. The given string must be exactly a single grapheme cluster, if not, None is returned.

pub fn as_str(&self) -> &str

Display this grapheme cluster as a string.

pub fn byte_len(&self) -> usize

Length in bytes of this grapheme cluster.

pub fn count_chars(&self) -> usize

Counts how many unicode characters compose this grapheme cluster.

pub fn has_diacritics(&self) -> bool

Returns whether this grapheme cluster has diacritics, i.e. additional characters.

pub fn strip_diacritics(&self) -> char

Returns the first character of this grapheme cluster, stripping diacritics.

pub fn is_char(&self) -> bool

Returns whether this grapheme cluster is a single character.

pub fn to_char(&self) -> Option<char>

Attempts to convert this grapheme cluster to a single character. Only succeeds if the cluster is composed of a single character.

pub fn is_alphabetic(&self) -> bool

Returns whether this grapheme cluster is alphabetic (possibly with diacritics).

pub fn is_alphabetic_char(&self) -> bool

Returns whether this grapheme cluster is alphabetic (and only if no diacritics are present).

pub fn is_ascii_alphabetic(&self) -> bool

Returns whether this grapheme cluster is a single ASCII alphabetic character.

pub fn is_numeric(&self) -> bool

Returns whether this grapheme cluster is numeric (possibly with diacritics).

pub fn is_numeric_char(&self) -> bool

Returns whether this grapheme cluster is numeric (and only if no diacritics are present).

pub fn is_ascii_numeric(&self) -> bool

Returns whether this grapheme cluster is a single ASCII numeric character.

pub fn is_alphanumeric(&self) -> bool

Returns whether this grapheme cluster is alphabetic or numeric (possibly with diacritics).

pub fn is_alphanumeric_char(&self) -> bool

Returns whether this grapheme cluster is alphabetic or numeric (and only if no diacritics are present).

pub fn is_ascii_alphanumeric(&self) -> bool

Returns whether this grapheme cluster is a single ASCII alphabetic or numeric character.

pub fn is_digit(&self, base: u32) -> bool

Returns whether this grapheme cluster is a digit of the given base. Only base 2 to 36 are allowed, using ASCII alphanumeric characters. Diacritics are allowed but ignored.

pub fn is_digit_char(&self, base: u32) -> bool

Returns whether this grapheme cluster is a digit of the given base. Only base 2 to 36 are allowed, using ASCII alphanumeric characters. Diacritics are not allowed and yield false.

pub fn to_digit(&self, base: u32) -> Option<u32>

Attempts to convert this grapheme cluster into a digit of the given base. Only base 2 to 36 are allowed, using ASCII alphanumeric characters. Diacritics are allowed but ignored.

pub fn char_to_digit(&self, base: u32) -> Option<u32>

Attempts to convert this grapheme cluster into a digit of the given base. Only base 2 to 36 are allowed, using ASCII alphanumeric characters. Diacritics are not allowed and yield None.

pub fn is_whitespace(&self) -> bool

Returns whether this grapheme cluster is whitespace (possibly with diacritics).

pub fn is_whitespace_char(&self) -> bool

Returns whether this grapheme cluster is whitespace (and only if no diacritics are present).