Trait bittongue::lexer::Lexer

pub trait Lexer {
    type TokenKind: TokenKind;
    fn generate_token(
        &mut self, 
        reader: &mut Reader, 
        diagnostics: &mut Diagnostics
    ) -> Result<Token<Self::TokenKind>, LexingError>;
}

A generic lexer.

Associated Types

type TokenKind: TokenKind

Token classes attached to tokens produced by this lexer.

Required methods

fn generate_token(
&mut self,
reader: &mut Reader,
diagnostics: &mut Diagnostics
) -> Result<Token<Self::TokenKind>, LexingError>

Attempts to generate a single Token. The lexer should manipulate the given Reader until a token is finished, and then, return the produced token. Ideally, final Reader’s position should not end up behind its initial position (regardig this method call), but if the lexer can handle not getting stuck, it should be fine.

Eventually, this method must produce a token whose kind is End-Of-File. More specifically, EOF token should be produced when Reader is in the end of input indeed. This method must not be called after EOF token is produced.

Although this method returns a Result, actual lexing errors should be inserted on Diagnostics via Diagnostics::raise. LexingError is just a semantic error for avoiding meaningless Options (i.e. better reading Result<Token<K>, LexingError> than Option<Token<K>>). Note that it is expected that when LexingError is returned, at least one error is placed on Diagnostics.