Module bittongue::source

This module provides utilities to handle source code of a programming language.

Example

use bittongue::source::Source;

let source = Source::new(
    "main.c",
    "int /* v́ */\nmain(int argc, char const *argv[])\n{\n    return 0;\n}\n"
);

let mut reader = source.reader();
reader.mark();
for _ in 0 .. 3 {
    assert!(reader.next());
}

let int_span = reader.span();
assert_eq!(&int_span.content(), "int");
assert_eq!(int_span.start().pos(), 0);
assert_eq!(int_span.start().line_column(), (0, 0));
assert_eq!(int_span.end().pos(), 3);
assert_eq!(int_span.end().line_column(), (0, 3));

for _ in 0 .. 9 {
    assert!(reader.next());
}
reader.mark();
for _ in 0 .. 4 {
    assert!(reader.next());
}

let main_span = reader.span();
assert_eq!(&main_span.content(), "main");
assert_eq!(main_span.start().pos(), 12);
assert_eq!(main_span.start().line_column(), (1, 0));
assert_eq!(main_span.end().pos(), 16);
assert_eq!(main_span.end().line_column(), (1, 4));

Structs

Location

The location in a source code.

NewlineIndices

Iterator over the newline indices of a source, in terms of grapheme clusters. Double-ended and sized.

Reader

A reader of a source code, a stream.

SegmentIndices

Iterator over the segment indices of a source, in terms of bytes. Double-ended and sized.

Source

A source code object, such as read from a file.

Span

A span (a range) in the source code.

SpanContent

A type that, when displayed, shows the span contents, rather than location.

Traits

SourceIndex

An index on a source code.