Struct bittongue::source::Source

pub struct Source { /* fields omitted */ }

A source code object, such as read from a file. Cloning this object results in simply incrementing a reference counter, thus sharing the source code object.

Implementations

impl Source

pub fn new<S0, S1>(name: S0, contents: S1) -> Self where
S0: Into<Box<str>>,
S1: Into<Box<str>>,

Creates a new source code object given its name and its contents.

Contents are rearranged as grapheme clusters.

pub fn name(&self) -> &str

The (file) name of the source.

pub fn len(&self) -> usize

The length the source.

pub fn contents(&self) -> &str

The contents of the source.

pub fn segments(&self) -> SegmentIndices<'_>

Iterator over the segment indices of the source, in terms of bytes.

pub fn newlines(&self) -> NewlineIndices<'_>

Iterator over the newline indices of the source, in terms of grapheme clusters.

pub fn get<I>(&self, indexer: I) -> Option<&I::Output> where
I: SourceIndex,

Indexes this source. It can be a single usize or a range of usize.

pub fn reader(&self) -> Reader

Creates a source code reader (a stream) from this source code object.